+++ types/storage.d.ts
@@ -1,0 +1,16 @@
+
+export interface Storage {
+  setUniversal(key: string, value: any, isJson?: boolean): string;
+  getUniversal(key: string, isJson?: boolean): any;
+  syncUniversal(key: string, defaultValue: any, isJson?: boolean): any;
+  // Local State
+  setState(key: string, val: any): string;
+  getState(key: string): string;
+  watchState(key: string, handler: (newValue: any) => void): any;
+  // Cookies
+  setCookie(key: string, val: any, options?: object): any;
+  getCookie(key: string, isJson?: boolean): any;
+  // Local Storage
+  setLocalStorage(key: string, val: any, isJson?: boolean): any;
+  getLocalStorage(key: string, isJson?: boolean): any;
+}
